refactor: bgColors 코드 위치 수정 및 주석제거
